feature: Quick expense and refactoring changes
