+++ split_expenses_ui/package.json
@@ -10,6 +10,7 @@
     "axios": "^0.21.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-dropzone": "^11.3.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.2",
     "sweetalert2": "^10.14.1",
