+++ split_expenses_ui/src/components/ProcessExpense/index.js
@@ -1,0 +1,1 @@
+export { default as ProcessExpense } from "./ProcessExpense";
