+++ split_expenses_ui/src/pages/Products/index.js
@@ -1,0 +1,2 @@
+import Products from './Products';
+export default Products;
