+++ split_expenses_ui/src/pages/QuickExpense/index.js
@@ -1,0 +1,2 @@
+import QuickExpense from './QuickExpense';
+export default QuickExpense;
