+++ split_expenses_ui/src/pages/Users/Users.jsx
@@ -1,33 +1,35 @@
 import React, { Component } from "react";
-import Modal from "./shared/modal";
-import { getProducts } from "../services/products";
+import { getusers } from "../../services";
+import Modal from "../../components/shared/modal";
 
-class Products extends Component {
+class Users extends Component {
   state = {
-    products: [],
+    users: [],
     isDisplayDialog: false,
   };
 
   componentDidMount = () => {
-    const fetchProducts = async () => {
-      const products = await getProducts();
-      console.log(products);
-      this.setState({ products });
+    const fetchUsers = async () => {
+      const users = await getusers();
+      this.setState({ users });
     };
-    fetchProducts();
+    fetchUsers();
   };
 
-  handleAddProduct = () => {
-    this.setState({ isDisplayDialog: true });
+  handleAddUser = () => {
+    console.log("Add user function clicked");
+    this.setState({
+      isDisplayDialog: true,
+    });
   };
 
   render() {
-    const { products, isDisplayDialog } = this.state;
+    const { users, isDisplayDialog } = this.state;
     return (
       <main>
         <div className="max-w-7xl mx-auto py-6 sm:px-6 lg:px-8">
           <button
-            onClick={this.handleAddProduct}
+            onClick={this.handleAddUser}
             className="hover:bg-light-blue-200 hover:text-light-blue-800 group flex items-center rounded-md bg-light-blue-100 text-light-blue-600 text-sm font-medium px-4 py-2"
           >
             <svg
@@ -56,19 +58,19 @@ class Products extends Component {
                             scope="col"
                             className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider"
                           >
-                            Product Name
+                            Name
                           </th>
                           <th
                             scope="col"
                             className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider"
                           >
-                            Quantity
+                            Email
                           </th>
                           <th
                             scope="col"
                             className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider"
                           >
-                            Total Price (Previous Total Price)
+                            Mobile Number
                           </th>
                           <th scope="col" className="relative px-6 py-3">
                             <span className="sr-only">Edit</span>
@@ -76,27 +78,27 @@ class Products extends Component {
                         </tr>
                       </thead>
                       <tbody className="bg-white divide-y divide-gray-200">
-                        {products &&
-                          products.length > 0 &&
-                          products.map((product) => {
+                        {users &&
+                          users.length > 0 &&
+                          users.map((user) => {
                             // console.log(user);
                             return (
                               <tr>
                                 <td className="px-6 py-4 whitespace-nowrap">
                                   <div className="flex items-center">
                                     <div className="text-sm font-medium text-gray-900">
-                                      {product.product_name}
+                                      {user.username}
                                     </div>
                                   </div>
                                 </td>
                                 <td className="px-6 py-4 whitespace-nowrap">
                                   <div className="text-sm text-gray-900">
-                                    {product.quantity}
+                                    {user.email}
                                   </div>
                                 </td>
                                 <td className="px-6 py-4 whitespace-nowrap">
                                   <div className="text-sm text-gray-900">
-                                    {product.price}
+                                    {user.phone_number}
                                   </div>
                                 </td>
                                 <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
@@ -124,4 +126,4 @@ class Products extends Component {
   }
 }
 
-export default Products;
+export default Users;
