+++ split_expenses_ui/src/pages/Users/index.js
@@ -1,0 +1,2 @@
+import Users from './Users';
+export default Users;
