+++ split_expenses_ui/src/services/index.js
@@ -1,0 +1,3 @@
+export * from './endpoints';
+export * from './products';
+export * from './users';
